add build cancellation functionality to prevent multiple builds piling up, and begin splitting out functions for qos handling
