+++ controllers/helpers.go
@@ -1,18 +1,22 @@
 package controllers
 
 import (
+	"context"
 	"crypto/sha1"
 	"crypto/sha256"
 	"encoding/base32"
 	"fmt"
 	"math/rand"
 	"regexp"
+	"sort"
 	"strconv"
 	"strings"
 	"time"
 
 	lagoonv1alpha1 "github.com/amazeeio/lagoon-kbd/api/v1alpha1"
+	"github.com/go-logr/logr"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
+	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
 const (
@@ -171,4 +175,35 @@ func variableExists(vars *[]LagoonEnvironmentVariable, name, value string) bool 
 		}
 	}
 	return exists
+}
+
+func cancelExtraBuilds(ctx context.Context, r client.Client, opLog logr.Logger, pendingBuilds *lagoonv1alpha1.LagoonBuildList, ns string, status string) error {
+	listOption := (&client.ListOptions{}).ApplyOptions([]client.ListOption{
+		client.InNamespace(ns),
+		client.MatchingLabels(map[string]string{"lagoon.sh/buildStatus": "Pending"}),
+	})
+	if err := r.List(ctx, pendingBuilds, listOption); err != nil {
+		return fmt.Errorf("Unable to list builds in the namespace, there may be none or something went wrong: %v", err)
+	}
+	opLog.Info(fmt.Sprintf("There are %v Pending builds", len(pendingBuilds.Items)))
+	// if we have any pending builds, then grab the latest one and make it running
+	// if there are any other pending builds, cancel them so only the latest one runs
+	sort.Slice(pendingBuilds.Items, func(i, j int) bool {
+		return pendingBuilds.Items[i].ObjectMeta.CreationTimestamp.After(pendingBuilds.Items[j].ObjectMeta.CreationTimestamp.Time)
+	})
+	if len(pendingBuilds.Items) > 0 {
+		for idx, pBuild := range pendingBuilds.Items {
+			pendingBuild := pBuild.DeepCopy()
+			if idx == 0 {
+				pendingBuild.Labels["lagoon.sh/buildStatus"] = status
+			} else {
+				// cancel any other pending builds
+				pendingBuild.Labels["lagoon.sh/buildStatus"] = "Cancelled"
+			}
+			if err := r.Update(ctx, pendingBuild); err != nil {
+				return err
+			}
+		}
+	}
+	return nil
 }
