+++ controllers/lagoonmonitor_controller.go
@@ -20,7 +20,6 @@ import (
 	"context"
 	"fmt"
 	"io"
-	"sort"
 
 	lagoonv1alpha1 "github.com/amazeeio/lagoon-kbd/api/v1alpha1"
 	"github.com/amazeeio/lagoon-kbd/handlers"
@@ -126,27 +125,7 @@ func (r *LagoonMonitorReconciler) Reconcile(req ctrl.Request) (ctrl.Result, erro
 		}
 		// if we have no running builds, then check for any pending builds
 		if len(runningBuilds.Items) == 0 {
-			listOption = (&client.ListOptions{}).ApplyOptions([]client.ListOption{
-				client.InNamespace(req.Namespace),
-				client.MatchingLabels(map[string]string{"lagoon.sh/buildStatus": "Pending"}),
-			})
-			if err := r.List(ctx, pendingBuilds, listOption); err != nil {
-				return ctrl.Result{}, fmt.Errorf("Unable to list builds in the namespace, there may be none or something went wrong: %v", err)
-			}
-			opLog.Info(fmt.Sprintf("There are %v Pending builds", len(pendingBuilds.Items)))
-			// if we have any pending builds, then grab the first one from the items as they are sorted by oldest pending first
-			if len(pendingBuilds.Items) > 0 {
-				// sort the pending builds by creation timestamp
-				sort.Slice(pendingBuilds.Items, func(i, j int) bool {
-					return pendingBuilds.Items[i].ObjectMeta.CreationTimestamp.Before(&pendingBuilds.Items[j].ObjectMeta.CreationTimestamp)
-				})
-				opLog.Info(fmt.Sprintf("Next build is: %s", pendingBuilds.Items[0].ObjectMeta.Name))
-				pendingBuild := pendingBuilds.Items[0].DeepCopy()
-				pendingBuild.Labels["lagoon.sh/buildStatus"] = "Running"
-				if err := r.Update(ctx, pendingBuild); err != nil {
-					return ctrl.Result{}, err
-				}
-			}
+			return ctrl.Result{}, cancelExtraBuilds(ctx, r, opLog, pendingBuilds, req.Namespace, "Running")
 		}
 	}
 	return ctrl.Result{}, nil
